feat: [US-001] - [Relocate reusable bento and widget-frame ownership]
