+++ docs/processes/development-guide-relevant-files.md
@@ -73,7 +73,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/src/App.stories.tsx` | Dashboard Storybook smoke surface | Current-selection and dashboard workflow stories should assert observable headings, dispatch markers, and mixed inference/script selected-work history output here so browser-backed regression coverage follows the customer-visible contract instead of helper structure. |
 | `ui/src/App.test.tsx` | Dashboard app-level regression surface | Mixed selected-work dispatch history coverage belongs here when a change needs full-app proof that inference-backed and script-backed workstation requests still share one `Workstation dispatches` surface. |
 | `ui/integration/event-stream-replay.integration.test.mjs` | Browser-backed dashboard integration harness | Keep browser-visible dashboard smoke, adjacent API stubs such as `/factory/~current` and cross-origin `POST /factory`, drag/drop `DataTransfer` file setup, and blob-download artifact capture in this one preview-backed Playwright/Vitest lane so UI integration coverage extends the built app without adding a second browser harness. |
-| `ui/src/components/ui/` | Shared website primitive layer | New reusable button, field, dialog, chart, table, skeleton, calendar, collapsible, and resizable primitives belong here instead of being reintroduced as feature-local dashboard chrome. |
+| `ui/src/components/ui/` | Shared website primitive layer | New reusable button, field, dialog, chart, table, skeleton, calendar, collapsible, resizable, and shared layout primitives such as bento boards or widget frames belong here; when legacy dashboard import paths still need compatibility, keep `components/dashboard/*` as thin re-export shims instead of the primary implementation owner. |
 | `ui/src/lib/cn.ts` | Shared class-composition utility | Shared website primitives should compose token-based Tailwind class strings through this utility instead of duplicating local class join helpers in each new primitive family. |
 | `ui/src/testing/replay-fixtures.ts` | Canonical replay fixture loader seam | Replay regression tests that need typed `FactoryEvent[]` or projected timeline snapshots should load checked-in `ui/integration/fixtures/*.jsonl` data through this helper instead of hand-parsing duplicate fixture maps in each test file. |
 | `ui/src/testing/replay-fixture-catalog.ts` | Replay and browser-workflow coverage contract | Browser-backed dashboard smoke coverage, including supplemental browser-only workflows that reuse replay fixtures such as PNG export/import roundtrips, should register scenario metadata here so coverage reporting and integration assertions stay on one source of truth. |
