+++ ui/src/components/dashboard/classnames.ts
@@ -1,3 +1,1 @@
-export function cx(...classes: Array<false | null | string | undefined>): string {
-  return classes.filter(Boolean).join(" ");
-}
+export * from "../ui/classnames";
