+++ ui/src/components/ui/classnames.ts
@@ -1,0 +1,3 @@
+export function cx(...classes: Array<false | null | string | undefined>): string {
+  return classes.filter(Boolean).join(" ");
+}
