+++ ui/src/components/ui/index.ts
@@ -1,12 +1,16 @@
+export * from "./agent-bento";
 export * from "./button";
 export * from "./calendar";
 export * from "./chart";
 export * from "./collapsible";
 export * from "./data-table";
+export * from "./dashboard-typography";
 export * from "./dialog";
 export * from "./input";
+export * from "./classnames";
 export * from "./resizable";
 export * from "./select";
 export * from "./skeleton";
 export * from "./table";
 export * from "./textarea";
+export * from "./widget-frame";
